requirements_no_torch.txt
absl-py==1.3.0
aliyundrive-webdav==1.10.1
asttokens
backcall
branca==0.5.0
brotlipy==0.7.0
cachetools==5.2.0
certifi
cffi
charset-normalizer
contourpy==1.0.6
countryinfo==0.1.2
cryptography
cycler==0.11.0
Cython==0.29.32
debugpy
decorator
entrypoints
executing
folium==0.13.0
fonttools==4.36.0
geographiclib==1.52
geopy==2.2.0
google-auth==2.15.0
google-auth-oauthlib==0.4.6
grpcio==1.51.1
idna
importlib-metadata==5.1.0
ipykernel
ipython
jedi
Jinja2==3.1.2
joblib==1.2.0
jupyter_client
jupyter_core
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline
nest-asyncio
networkx==2.8.8
numexpr==2.8.4
numpy==1.25.2
oauthlib==3.2.2
packaging==21.3
pandas==1.3.5
parso
pexpect
pickleshare
Pillow==9.2.0
prompt-toolkit
protobuf==3.20.3
psutil
ptyprocess
pure-eval
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser
Pygments
pyOpenSSL
pyparsing==3.0.9
pyshp==2.3.1
PySocks
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
pyzmq
requests
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.1.3
scipy==1.9.3
seaborn==0.12.1
setproctitle==1.3.2
shapely==2.0.0
six
stack-data
tables==3.7.0
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
torchinfo==1.7.1
tornado
tqdm==4.64.1
traitlets
typing_extensions
urllib3
wcwidth
Werkzeug==2.2.2
zipp==3.11.0
